First working of Sunrise Workbench + Kuka PRC XML loading
POST SETUP OF SUNRISE WORKBENCH (this was easy on a new desktop as the kuka lbr iiwa 14 is already configured. Sunrise workbench can just open a new project and download the current configuration). 

recent fixes:

1. error with import com.kuka.task.ITaskLogger
Commented out and replaced with
iApplicationUI.displayModalDialog instead for debugging

2. NullPointerException
issue with the TCP frame definition. fixed in PRC_CORE.java file by defining same as the robot's flange (works for now it seems, but won't be appropriate when an actual EEF is placed)? 

3. PRC XML file generated
go to /KUKA_PRC_XML/notes.txt to understand changes